Task-9/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
python-dotenv>=0.19.0
pydantic>=1.8.0
python-multipart>=0.0.5

assemblyai
google-generativeai
